requirements.txt
Django==2.0.3
django-ckeditor==5.4.0
django-crispy-forms==1.7.1
django-js-asset==1.1.0
django-registration-redux==2.5
Markdown==2.6.11
Pillow==5.0.0
pytz==2018.3
